+++ src/components/GaragesContainer.js
@@ -2,13 +2,13 @@ import React from 'react';
 import GarageContainer from './GarageContainer'
 import Row from 'react-bootstrap/Row'
 import Container from 'react-bootstrap/Container'
-import { cookieName } from '../constants/Cookie'
-import Cookies from 'universal-cookie'
+import Col from 'react-bootstrap/Col'
 import GETGaragebyUserID from './DB Connection/GETGaragebyUserID'
 import GETGarages from './DB Connection/GETGarages';
 import GETGaragesFiltered from './DB Connection/GETGaragesFiltered'
 import TransformGarageData from './Transform/TransformGarageData'
 import LoadingIndicator from './LoadingIndicator'
+import RecuperarCookie from './RecuperarCookie'
 import '../styles/garagecomp.css'
 
 class GaragesContainer extends React.Component {
@@ -22,9 +22,9 @@ class GaragesContainer extends React.Component {
     }
   }
 
-  cookies = new Cookies();
-  cookie = this.cookies.get(cookieName);
+  cookie = RecuperarCookie();
   user_id = { "user_id": this.cookie.user_id };
+  rol = this.cookie.rol;
 
   async componentDidMount() {
 
@@ -79,33 +79,69 @@ class GaragesContainer extends React.Component {
       if (this.state.loaded === false) {
         this.setState({ ...this.state, "loaded": true })
       }
-      return (
-        <Container fluid>
-          <Row className="ml-md-5 mr-md-5 justify-content-between">
-            <LoadingIndicator />
-            {
-              garages.map((garage) => {
-                const transformed_data = TransformGarageData(garage, this.cookie.rol);
-                return (<GarageContainer garage_data={transformed_data} />)
-              })
-            }
-          </Row>
-        </Container>
-      )
+      if (garages.length % 2 !== 0) {
+        return (
+          <Container fluid>
+            <Row className="ml-md-5 mr-md-5 justify-content-around">
+              <LoadingIndicator />
+              {
+                garages.map((garage) => {
+                  const transformed_data = TransformGarageData(garage, this.cookie.rol);
+                  return (<GarageContainer garage_data={transformed_data} />)
+                })
+              }
+              <Col className="mr-md-2 mt-4 garagecomp invisible" lg={5} ></Col>
+            </Row>
+          </Container>
+        )
+      } else {
+        return (
+          <Container fluid>
+            <Row className="ml-md-5 mr-md-5 justify-content-around">
+              <LoadingIndicator />
+              {
+                garages.map((garage) => {
+                  const transformed_data = TransformGarageData(garage, this.cookie.rol);
+                  return (<GarageContainer garage_data={transformed_data} />)
+                })
+              }
+            </Row>
+          </Container>
+        )
+      }
+
     }
     else {
       if (this.state.loaded === false) {
-        return (<LoadingIndicator />)
+        return (<>
+          {(parseInt(this.rol, 10) === 2) &&
+            <Row className="mt-3 garagecomp lg={5} mx-auto">
+              <div>
+                <h4>Todavía no registraste garages.</h4>
+              </div>
+            </Row>
+          }
+          <LoadingIndicator />
+        </>)
       }
       else {
         return (
           <>
             <LoadingIndicator />
-            <Row className="ml-5 mr-5 mt-3 garagecomp lg={5} justify-content-between">
-              <div>
-                <h4>El filtro seleccionado no ha arrojado resultados, inténtelo nuevamente.</h4>
-              </div>
-            </Row>
+            {(parseInt(this.rol, 10) === 1) &&
+              <Row className="mt-3 garagecomp lg={5} mx-auto">
+                <div>
+                  <h4>El filtro seleccionado no ha arrojado resultados, inténtelo nuevamente.</h4>
+                </div>
+              </Row>
+            }
+            {(parseInt(this.rol, 10) === 2) &&
+              <Row className="mt-3 garagecomp lg={5} mx-auto">
+                <div>
+                  <h4>Todavía no registraste garages.</h4>
+                </div>
+              </Row>
+            }
           </>
         )
       }