+++ src/components/IndexContainer.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import Header from './Header'
 import ClienteNavBar from './ClienteNavBar'
 import ChequearCookie from './ChequearCookie'
 import GaragesContainer from './GaragesContainer'
@@ -7,23 +6,21 @@ import UserContainer from './UserContainer'
 
 const IndexContainer = () => {
   const salidaUser =
-  <section>
-    <Header />
-    <UserContainer/>
-  </section>
+    <section>
+      <UserContainer />
+    </section>
 
   const salidaCliente =
-  <section>
-    <Header />
-    <ClienteNavBar />
-    <GaragesContainer />
-  </section>
-  
-  if(window.location.pathname === '/index'){
+    <section>
+      <ClienteNavBar />
+      <GaragesContainer />
+    </section>
+
+  if (window.location.pathname === '/index') {
     return (ChequearCookie(salidaUser, '/index', '/clientindex', '/login'));
   }
-  
-  if(window.location.pathname === '/clientindex'){
+
+  if (window.location.pathname === '/clientindex') {
     return (ChequearCookie(salidaCliente, '/index', '/clientindex', '/login'));
   }