+++ src/components/UserNavBar.js
@@ -1,0 +1,20 @@
+import React from 'react'
+import Button from 'react-bootstrap/Button'
+import Container from 'react-bootstrap/Container'
+import { Link } from 'react-router-dom'
+import '../styles/ClienteNavBar.css'
+
+
+const UserNavBar = (props) => {
+  return (
+    <div className="cliente-navbar">
+      <Container className="client-nav-wrapper">
+        <Link to='/index'>
+          <Button className="search-btn" variant="info">Buscar Garages</Button>
+        </Link>
+      </Container>
+    </div>
+  )
+}
+
+export default UserNavBar