+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 **Notas de Versión**
 
+### Versión 0.4 - Release 4
+
+**03/07/2020**
+
+### Cambios:
+
+#### Añadido:
+- Vista de reservas para clientes y usuarios
+- Reservar un lugar para un usuario
+- Recuperar datos de reservas por usuario
+- Recuperar datos de reservas por garage para clientes
+- Encriptación de Cookie para seguridad mejorada
+- Los datos de Login son ahora requeridos
+- Cambios estéticos:
+	- Logo
+	- Background
+
 ### Versión 0.3 - Release 3
 
 **18/06/2020**
@@ -36,4 +53,3 @@
 - Login completo
 - Cookie de sesión para evitar acceso a áreas donde se requiera iniciar sesión, y para usar autologin
 - Link a redes edes sociales
-